+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.6] - Unreleased
+## [1.0.0] - Unreleased
+### Changed
+- How it works
+
+### Added
+- Connection, RedoxClient, Authorization
+- Global configuration block
+- Gems httparty, hashie
+
+### Removed
+- Identifiers, Demographics
+
+## [0.1.6] - 2019-04-22
 ### Added
 - Meta model
 - Response and Authentication classes
@@ -58,7 +70,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/WeInfuse/redox/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/WeInfuse/redox/compare/v0.1.6...HEAD
+[0.1.6]: https://github.com/WeInfuse/redox/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/WeInfuse/redox/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/WeInfuse/redox/compare/0.1.3...v0.1.4
 [0.1.3]: https://github.com/WeInfuse/redox/compare/0.1.2...0.1.3